# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SahabatBarokahAI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterRequest {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export type UserProfilePlan =
  (typeof UserProfilePlan)[keyof typeof UserProfilePlan];

export const UserProfilePlan = {
  free: "free",
  pro: "pro",
  business: "business",
} as const;

export type UserProfileRole =
  (typeof UserProfileRole)[keyof typeof UserProfileRole];

export const UserProfileRole = {
  user: "user",
  admin: "admin",
} as const;

export interface UserProfile {
  id: number;
  name: string;
  email: string;
  plan: UserProfilePlan;
  role: UserProfileRole;
  createdAt: string;
  dailyUsage: number;
}

export interface AuthResponse {
  user: UserProfile;
  token: string;
}

export interface StrategyRequest {
  businessType: string;
  targetMarket: string;
  budget: string;
}

export interface MenuRequest {
  foodType: string;
  targetCustomer: string;
  priceRange: string;
}

export interface ProfitRequest {
  productName: string;
  costPerItem: number;
  sellingPrice: number;
  estimatedDailySales: number;
}

export interface MarketingRequest {
  productName: string;
  promotionType: string;
  targetAudience: string;
}

export type GenerationResultInput = { [key: string]: unknown };

export interface GenerationResult {
  id: number;
  tool: string;
  input: GenerationResultInput;
  result: string;
  createdAt: string;
}

export type GenerationInput = { [key: string]: unknown };

export interface Generation {
  id: number;
  tool: string;
  input: GenerationInput;
  result: string;
  createdAt: string;
}

export interface AdminStats {
  totalUsers: number;
  totalGenerations: number;
  todayGenerations: number;
  freeUsers: number;
  proUsers: number;
  businessUsers: number;
}

export type AdminUserPlan = (typeof AdminUserPlan)[keyof typeof AdminUserPlan];

export const AdminUserPlan = {
  free: "free",
  pro: "pro",
  business: "business",
} as const;

export type AdminUserRole = (typeof AdminUserRole)[keyof typeof AdminUserRole];

export const AdminUserRole = {
  user: "user",
  admin: "admin",
} as const;

export interface AdminUser {
  id: number;
  name: string;
  email: string;
  plan: AdminUserPlan;
  role: AdminUserRole;
  createdAt: string;
  totalGenerations: number;
}

export type UpdatePlanRequestPlan =
  (typeof UpdatePlanRequestPlan)[keyof typeof UpdatePlanRequestPlan];

export const UpdatePlanRequestPlan = {
  free: "free",
  pro: "pro",
  business: "business",
} as const;

export interface UpdatePlanRequest {
  plan: UpdatePlanRequestPlan;
}

export interface DailyUsagePoint {
  date: string;
  count: number;
}

export interface ToolUsagePoint {
  tool: string;
  count: number;
}

export interface UserGrowthPoint {
  date: string;
  count: number;
}

export interface AnalyticsData {
  dailyUsage: DailyUsagePoint[];
  toolUsage: ToolUsagePoint[];
  userGrowth: UserGrowthPoint[];
}