# lib/api-zod/src/generated/types/analyticsData.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SahabatBarokahAI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DailyUsagePoint } from "./dailyUsagePoint";
import type { ToolUsagePoint } from "./toolUsagePoint";
import type { UserGrowthPoint } from "./userGrowthPoint";

export interface AnalyticsData {
  dailyUsage: DailyUsagePoint[];
  toolUsage: ToolUsagePoint[];
  userGrowth: UserGrowthPoint[];
}