# lib/api-zod/src/generated/types/generationResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SahabatBarokahAI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GenerationResultInput } from "./generationResultInput";

export interface GenerationResult {
  id: number;
  tool: string;
  input: GenerationResultInput;
  result: string;
  createdAt: Date;
}