# lib/api-zod/src/generated/types/menuRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SahabatBarokahAI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MenuRequest {
  foodType: string;
  targetCustomer: string;
  priceRange: string;
}