# lib/api-zod/src/generated/types/strategyRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SahabatBarokahAI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StrategyRequest {
  businessType: string;
  targetMarket: string;
  budget: string;
}