# lib/api-zod/src/generated/types/updatePlanRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SahabatBarokahAI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdatePlanRequestPlan } from "./updatePlanRequestPlan";

export interface UpdatePlanRequest {
  plan: UpdatePlanRequestPlan;
}